+++ CHANGELOG.md
@@ -7,7 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [0.4.0] - 2026-Apr-2
 - `default_rules` adjustments for v6 NDP and DHCPv4/6
 - stricter RLIMIT validations
-- refactor of all packages
+- strengthened .toml string validations
 
 ## [0.3.5] - 2026-Apr-26
 - major refactor of all packages